Revert "moved execution logic to test module __init__.py"
This reverts commit abd3afb3367ac07aa3e6e23d37c68d6a86432415.
